Unify test case instantiator
