+++ tests/generators/fork_choice_generated/instantiators/helpers.py
@@ -1,4 +1,4 @@
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from .debug_helpers import print_epoch
 from eth2spec.test.helpers.state import (
     next_slot,
@@ -10,7 +10,14 @@ from eth2spec.test.helpers.block import (
     build_empty_block,
     sign_block,
 )
-
+from eth2spec.test.helpers.fork_choice import (
+    on_tick_and_append_step,
+    add_attestation,
+    add_attester_slashing,
+    add_block,
+    output_store_checks,
+)
+from .debug_helpers import print_head
 
 @dataclass
 class ProtocolMessage:
@@ -24,8 +31,9 @@ class FCTestData:
     anchor_block: object
     anchor_state: object
     blocks: list[ProtocolMessage]
-    atts: list[ProtocolMessage]
-    slashings: list[ProtocolMessage]
+    atts: list[ProtocolMessage] = field(default_factory=list)
+    slashings: list[ProtocolMessage] = field(default_factory=list)
+    store_final_time: int = 0
 
 
 class BranchTip:
@@ -216,3 +224,70 @@ def advance_state_to_anchor_epoch(spec, state, anchor_epoch, debug) -> ([], Bran
     return signed_blocks, anchor_tip
 
 
+def _on_tick_and_append_step(spec, store, slot, test_steps):
+    time = slot * spec.config.SECONDS_PER_SLOT + store.genesis_time
+    on_tick_and_append_step(spec, store, time, test_steps)
+    assert store.time == time
+
+
+def yield_fork_choice_test_case(spec, store, test_data: FCTestData, debug: bool):
+    # Yield meta
+    for k, v in test_data.meta.items():
+        yield k, 'meta', v
+
+    # Yield anchor state and block initialization
+    yield 'anchor_state', test_data.anchor_state
+    yield 'anchor_block', test_data.anchor_block
+
+    test_steps = []
+    _on_tick_and_append_step(spec, store, test_data.anchor_state.slot, test_steps)
+
+    # Apply generated messages
+    max_block_slot = max(b.payload.message.slot for b in test_data.blocks)
+    max_attestation_slot = max(a.payload.data.slot for a in test_data.atts) + 1 if any(test_data.atts) else 0
+    max_slashing_slot = max(max(s.payload.attestation_1.data.slot,
+                                s.payload.attestation_2.data.slot)
+                            for s in test_data.slashings) + 1 if any(test_data.slashings) else 0
+
+    start_slot = min(b.payload.message.slot for b in test_data.blocks)
+    end_slot = max(max_block_slot, max_attestation_slot, max_slashing_slot)
+
+    # Advance time to start_slot
+    _on_tick_and_append_step(spec, store, start_slot, test_steps)
+
+    # Apply messages to store
+    for slot in range(start_slot, end_slot + 1):
+        # on_tick
+        _on_tick_and_append_step(spec, store, slot, test_steps)
+
+        # on_attestation for attestations from the previous slot
+        for attestation_message in (a for a in test_data.atts if a.payload.data.slot == slot - 1):
+            yield from add_attestation(spec, store, attestation_message.payload,
+                                       test_steps, valid=attestation_message.valid)
+
+        # on_attester_slashing for slashing from the previous slot
+        for attester_slashing_message in (s for s in test_data.slashings
+                                          if max(s.payload.attestation_1.data.slot,
+                                                 s.payload.attestation_2.data.slot) == slot - 1):
+            yield from add_attester_slashing(spec, store, attester_slashing_message.payload,
+                                             test_steps, valid=attester_slashing_message.valid)
+
+        # on_block for blocks from the current slot
+        for signed_block_message in (b for b in test_data.blocks if b.payload.message.slot == slot):
+            yield from add_block(spec, store, signed_block_message.payload, test_steps, signed_block_message.valid)
+
+            block_root = signed_block_message.payload.message.hash_tree_root()
+            if signed_block_message.valid:
+                assert store.blocks[block_root] == signed_block_message.payload.message
+            else:
+                assert block_root not in store.blocks.values()
+
+    if store.time < test_data.store_final_time:
+        on_tick_and_append_step(spec, store, test_data.store_final_time, test_steps)
+
+    if debug:
+        print('               head: ' + print_head(spec, store))
+
+    output_store_checks(spec, store, test_steps, with_filtered_block_weights=True)
+
+    yield 'steps', test_steps
